change Color from [f32; 3] to [u8; 3]

+++ src/components.rs
@@ -152,6 +152,6 @@ pub struct RenderOnMap;
 
 pub struct Renderable {
     pub ch: char,
-    pub fg: [f32; 3],
-    pub bg: [f32; 3],
+    pub fg: [u8; 3],
+    pub bg: [u8; 3],
 }
